refactor: [165] fixing errors
