+++ fenn/cli/__init__.py
@@ -1,11 +1,9 @@
 import argparse
 
-import fenn.cli.auth as auth
 import fenn.cli.dashboard as dashboard
 import fenn.cli.grid as grid
 import fenn.cli.list as list
 import fenn.cli.pull as pull
-import fenn.cli.run as run
 
 
 def build_parser() -> argparse.ArgumentParser:
@@ -60,90 +58,6 @@ def build_parser() -> argparse.ArgumentParser:
     )
     p_dash.add_argument("--debug", action="store_true", help="Run in debug mode")
     p_dash.set_defaults(func=dashboard.execute)
-
-    # ========= RUN =========
-    p_run = subparsers.add_parser(
-        "run",
-        help="Run a Fenn project on the Fenn remote service",
-    )
-    p_run.add_argument(
-        "script",
-        nargs="?",
-        default=None,
-        help="Path to the entrypoint script (default: main.py)",
-    )
-    p_run.add_argument(
-        "--api-key",
-        default=None,
-        help="API key (overrides env, credentials file, and .env)",
-    )
-    p_run.add_argument(
-        "--profile",
-        default=None,
-        help="Credentials profile name (default: 'default' or $FENN_PROFILE)",
-    )
-    p_run.add_argument(
-        "--max-runtime",
-        type=int,
-        default=10,
-        help="Maximum allowed wall-time in minutes (server enforces; default: 10)",
-    )
-    p_run.add_argument(
-        "--detach",
-        action="store_true",
-        help="Submit the job and exit without streaming logs",
-    )
-    p_run.add_argument(
-        "--no-download",
-        action="store_true",
-        help="Do not download artifacts on completion",
-    )
-    p_run.add_argument(
-        "--include",
-        action="append",
-        metavar="PATH",
-        help="Extra path (relative to CWD) to include in the upload tarball",
-    )
-    p_run.add_argument(
-        "--exclude",
-        action="append",
-        metavar="PATTERN",
-        help="Extra shell-glob pattern to exclude from the upload tarball",
-    )
-    p_run.set_defaults(func=run.execute)
-
-    # ========= AUTH =========
-    p_auth = subparsers.add_parser(
-        "auth", help="Manage credentials for the Fenn remote service"
-    )
-    auth_subparsers = p_auth.add_subparsers(dest="auth_command", required=True)
-
-    p_login = auth_subparsers.add_parser("login", help="Save an API key for a profile")
-    p_login.add_argument(
-        "--profile", default=None, help="Profile name (default: 'default')"
-    )
-    p_login.add_argument(
-        "--api-key",
-        default=None,
-        help="API key (if omitted, will prompt or read from stdin)",
-    )
-    p_login.set_defaults(func=auth.execute)
-
-    p_status = auth_subparsers.add_parser(
-        "status", help="Show the currently configured profile and credit balance"
-    )
-    p_status.add_argument(
-        "--profile", default=None, help="Profile name (default: 'default')"
-    )
-    p_status.set_defaults(func=auth.execute)
-
-    p_logout = auth_subparsers.add_parser(
-        "logout", help="Remove a profile from the credentials file"
-    )
-    p_logout.add_argument(
-        "--profile", default=None, help="Profile name (default: 'default')"
-    )
-    p_logout.set_defaults(func=auth.execute)
 
     # ========= GRID =========
